Refactor BucketInfo package location
