+++ src/main/java/org/amoseman/budgetingbackend/resource/BucketResource.java
@@ -10,7 +10,7 @@ import org.amoseman.budgetingbackend.exception.BucketAlreadyExistsException;
 import org.amoseman.budgetingbackend.exception.BucketDoesNotExistException;
 import org.amoseman.budgetingbackend.exception.TotalBucketShareExceededException;
 import org.amoseman.budgetingbackend.pojo.bucket.Bucket;
-import org.amoseman.budgetingbackend.pojo.bucket.op.BucketInfo;
+import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
 import org.amoseman.budgetingbackend.service.BucketService;
 
 import java.util.List;
