+++ src/test/java/org/amoseman/budgetingbackend/dao/impl/sql/BucketDAOImplTest.java
@@ -10,7 +10,7 @@ import org.amoseman.budgetingbackend.exception.BucketDoesNotExistException;
 import org.amoseman.budgetingbackend.exception.UserAlreadyExistsException;
 import org.amoseman.budgetingbackend.pojo.account.op.CreateAccount;
 import org.amoseman.budgetingbackend.pojo.bucket.Bucket;
-import org.amoseman.budgetingbackend.pojo.bucket.op.BucketInfo;
+import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
 import org.amoseman.budgetingbackend.service.AccountService;
 import org.amoseman.budgetingbackend.util.Now;
 import org.jooq.DSLContext;
