+++ src/test/java/org/amoseman/budgetingbackend/dao/impl/sql/FinanceRecordDAOImplTest.java
@@ -8,7 +8,7 @@ import org.amoseman.budgetingbackend.database.impl.sql.sqlite.DatabaseConnection
 import org.amoseman.budgetingbackend.exception.*;
 import org.amoseman.budgetingbackend.pojo.TimeRange;
 import org.amoseman.budgetingbackend.pojo.account.op.CreateAccount;
-import org.amoseman.budgetingbackend.pojo.bucket.op.BucketInfo;
+import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
 import org.amoseman.budgetingbackend.pojo.record.Expense;
 import org.amoseman.budgetingbackend.pojo.record.Income;
 import org.amoseman.budgetingbackend.pojo.record.info.ExpenseInfo;
