+++ src/test/java/org/amoseman/budgetingbackend/service/BucketServiceTest.java
@@ -12,7 +12,7 @@ import org.amoseman.budgetingbackend.database.impl.sql.sqlite.DatabaseConnection
 import org.amoseman.budgetingbackend.exception.*;
 import org.amoseman.budgetingbackend.pojo.account.op.CreateAccount;
 import org.amoseman.budgetingbackend.pojo.bucket.Bucket;
-import org.amoseman.budgetingbackend.pojo.bucket.op.BucketInfo;
+import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
 import org.amoseman.budgetingbackend.pojo.record.info.ExpenseInfo;
 import org.amoseman.budgetingbackend.pojo.record.info.IncomeInfo;
 import org.jooq.DSLContext;
