+++ src/test/java/org/amoseman/budgetingbackend/service/FinanceRecordServiceTest.java
@@ -10,7 +10,7 @@ import org.amoseman.budgetingbackend.database.DatabaseConnection;
 import org.amoseman.budgetingbackend.database.impl.sql.sqlite.DatabaseConnectionImpl;
 import org.amoseman.budgetingbackend.exception.*;
 import org.amoseman.budgetingbackend.pojo.account.op.CreateAccount;
-import org.amoseman.budgetingbackend.pojo.bucket.op.BucketInfo;
+import org.amoseman.budgetingbackend.pojo.bucket.BucketInfo;
 import org.amoseman.budgetingbackend.pojo.record.Expense;
 import org.amoseman.budgetingbackend.pojo.record.Income;
 import org.amoseman.budgetingbackend.pojo.record.info.ExpenseInfo;
